toyDatabase/requirements.txt
faiss-cpu
numpy
matplotlib
sentence-transformers
torch
transformers
